ADD INLINE DETOUR!
The exception detour is powerful and cross-platform, but it is too slow
and requires a lot of cpu cycles. If we detour a hot function, it will
eat up all the cpu time.
With the inline detour, all plugins can increase the performance while
don't need to change any code due to our design.
genkai used to take all the cpu time in my laptop(thanks god it is
single thread program), now it drops to about 0-2%.
For now we only support x86 with the inline detour. This method has a
lot of restrictions. If a function cannot be detoured by inline, we will
use the exception detour.